# slab/src/main/webapp/openapi/slab.yml
openapi: '3.0.2'
info:
  title: slab
  version: '1.0'
servers:
  - url: http://localhost:8080/slab/api/v1
...
paths:
...
  /fotos:
    post:
      tags:
        - fotos
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Rocodromo'
        required: true
      responses:
        '201':
          description: CREATED
          content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Rocodromo'
    get:
      tags:
      - fotos
      parameters:
        - name: id 
          in: path
          required: true
          schema: 
            type: integer
            format: int64
      responses:
          '200':
            description: OK
            content:
              application/json:
                  schema:
                    $ref: '#/components/schemas/Rocodromo'
...
components:
  schemas:
...
    Rocodromo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        ubicacion:
          type: string
...